updated update-nvd-api
